Refactor routing directory structure
