+++ routes/api/auth-routes.js
@@ -1,15 +1,15 @@
 const express = require('express');
 const router = express.Router();
 const { check, oneOf } = require('express-validator');
-const { protect } = require('../middleware/auth');
-const { ALPHA_WHITESPACE_REGEX } = require('../utils/regex');
+const { protect } = require('../../middleware/auth');
+const { ALPHA_WHITESPACE_REGEX } = require('../../utils/regex');
 const {
   INVALID_EMAIL_MSG,
   INVALID_ALPHA_SPACE_MSG,
   INVALID_PASSWORD_MSG,
   NO_FIELD_UPDATED_MSG,
   INVALID_EXISTING_MSG,
-} = require('../utils/inputExceptionMsg');
+} = require('../../utils/inputExceptionMsg');
 
 // import controllers here
 const {
@@ -22,8 +22,8 @@ const {
   confirmEmail,
   forgetPassword,
   resetPassword,
-} = require('../controllers/auth');
-const { checkInputError } = require('../middleware/inputValidation');
+} = require('../../controllers/auth');
+const { checkInputError } = require('../../middleware/inputValidation');
 
 // map routes to controller
 router.get('/logout', protect, logout);
