+++ routes/api/faq-routes.js
@@ -1,13 +1,13 @@
 const express = require('express');
 const router = express.Router({ mergeParams: true });
 const { check, oneOf } = require('express-validator');
-const advancedResults = require('../middleware/advancedResults');
-const { protect, authorise } = require('../middleware/auth');
-const { checkInputError } = require('../middleware/inputValidation');
+const advancedResults = require('../../middleware/advancedResults');
+const { protect, authorise } = require('../../middleware/auth');
+const { checkInputError } = require('../../middleware/inputValidation');
 const {
   NO_FIELD_UPDATED_MSG,
   INVALID_FIELD_MSG,
-} = require('../utils/inputExceptionMsg');
+} = require('../../utils/inputExceptionMsg');
 
 // import controllers here
 const {
@@ -16,7 +16,7 @@ const {
   createFaq,
   updateFaq,
   deleteFaq,
-} = require('../controllers/faqs');
+} = require('../../controllers/faqs');
 
 router.route('/').get(advancedResults('faqs'), getFaqs);
 router.route('/:id').get(getFaq);
