+++ routes/api/hashtag-routes.js
@@ -1,13 +1,13 @@
 const express = require('express');
 const router = express.Router({ mergeParams: true });
 const { check, oneOf } = require('express-validator');
-const advancedResults = require('../middleware/advancedResults');
-const { protect, authorise } = require('../middleware/auth');
-const { checkInputError } = require('../middleware/inputValidation');
+const advancedResults = require('../../middleware/advancedResults');
+const { protect, authorise } = require('../../middleware/auth');
+const { checkInputError } = require('../../middleware/inputValidation');
 const {
   NO_FIELD_UPDATED_MSG,
   INVALID_FIELD_MSG,
-} = require('../utils/inputExceptionMsg');
+} = require('../../utils/inputExceptionMsg');
 
 // import controllers here
 const {
@@ -16,8 +16,8 @@ const {
   createHashtag,
   updateHashtag,
   deleteHashtag,
-} = require('../controllers/hashtags');
-const { HASHTAG_REGEX } = require('../utils/regex');
+} = require('../../controllers/hashtags');
+const { HASHTAG_REGEX } = require('../../utils/regex');
 
 router.route('/').get(advancedResults('hashtags'), getHashtags);
 router.route('/:id').get(getHashtag);
