+++ routes/api/like-routes.js
@@ -1,13 +1,18 @@
 const express = require('express');
 const router = express.Router({ mergeParams: true });
 const { check } = require('express-validator');
-const advancedResults = require('../middleware/advancedResults');
-const { protect } = require('../middleware/auth');
-const { checkInputError } = require('../middleware/inputValidation');
-const { INVALID_FIELD_MSG } = require('../utils/inputExceptionMsg');
+const advancedResults = require('../../middleware/advancedResults');
+const { protect } = require('../../middleware/auth');
+const { checkInputError } = require('../../middleware/inputValidation');
+const { INVALID_FIELD_MSG } = require('../../utils/inputExceptionMsg');
 
 // import controllers here
-const { getLikes, getLike, newLike, unLike } = require('../controllers/likes');
+const {
+  getLikes,
+  getLike,
+  newLike,
+  unLike,
+} = require('../../controllers/likes');
 
 router.route('/').get(advancedResults('likes'), getLikes);
 router.route('/:like_id').get(getLike);
