+++ routes/api/listingstory-routes.js
@@ -1,20 +1,20 @@
 const express = require('express');
 const router = express.Router();
 const { check, oneOf } = require('express-validator');
-const advancedResults = require('../middleware/advancedResults');
-const { protect, authorise } = require('../middleware/auth');
-const { checkInputError } = require('../middleware/inputValidation');
+const advancedResults = require('../../middleware/advancedResults');
+const { protect, authorise } = require('../../middleware/auth');
+const { checkInputError } = require('../../middleware/inputValidation');
 const {
   NO_FIELD_UPDATED_MSG,
   INVALID_FIELD_MSG,
-} = require('../utils/inputExceptionMsg');
+} = require('../../utils/inputExceptionMsg');
 
 // import controllers here
 const {
   getListingStories,
   getListingStory,
   updateListingStory,
-} = require('../controllers/listingstories');
+} = require('../../controllers/listingstories');
 
 // map routes to controller
 router.route('/').get(advancedResults('listingstories'), getListingStories);
