+++ routes/api/profile-routes.js
@@ -2,15 +2,15 @@ const express = require('express');
 const router = express.Router();
 const { check, oneOf } = require('express-validator');
 
-const advancedResults = require('../middleware/advancedResults');
-const { protect, authorise } = require('../middleware/auth');
-const { checkInputError } = require('../middleware/inputValidation');
+const advancedResults = require('../../middleware/advancedResults');
+const { protect, authorise } = require('../../middleware/auth');
+const { checkInputError } = require('../../middleware/inputValidation');
 const {
   INVALID_TIMESTAMP_MSG,
   INVALID_FIELD_MSG,
-} = require('../utils/inputExceptionMsg');
-const { DATETIME_REGEX } = require('../utils/regex');
-const { uploadFile } = require('../utils/fileUploader');
+} = require('../../utils/inputExceptionMsg');
+const { DATETIME_REGEX } = require('../../utils/regex');
+const { uploadFile } = require('../../utils/fileUploader');
 
 // Include other resource's controllers to access their endpoints
 const likeRoute = require('./like-routes');
@@ -30,8 +30,8 @@ const {
   updateProfile,
   verifyProfile,
   uploadPic,
-} = require('../controllers/profiles');
-const { NO_FIELD_UPDATED_MSG } = require('../utils/inputExceptionMsg');
+} = require('../../controllers/profiles');
+const { NO_FIELD_UPDATED_MSG } = require('../../utils/inputExceptionMsg');
 
 // map routes to controller
 router.route('/').get(advancedResults('profiles'), getProfiles);
