+++ routes/api/skill-routes.js
@@ -1,14 +1,14 @@
 const express = require('express');
 const router = express.Router({ mergeParams: true });
 const { check } = require('express-validator');
-const advancedResults = require('../middleware/advancedResults');
-const { protect, authorise } = require('../middleware/auth');
-const { checkInputError } = require('../middleware/inputValidation');
-const { ALPHA_WHITESPACE_REGEX } = require('../utils/regex');
+const advancedResults = require('../../middleware/advancedResults');
+const { protect, authorise } = require('../../middleware/auth');
+const { checkInputError } = require('../../middleware/inputValidation');
+const { ALPHA_WHITESPACE_REGEX } = require('../../utils/regex');
 const {
   INVALID_ALPHA_SPACE_MSG,
   NO_FIELD_UPDATED_MSG,
-} = require('../utils/inputExceptionMsg');
+} = require('../../utils/inputExceptionMsg');
 
 // import controllers here
 const {
@@ -17,7 +17,7 @@ const {
   createSkill,
   updateSkill,
   deleteSkill,
-} = require('../controllers/skills');
+} = require('../../controllers/skills');
 
 router.route('/').get(advancedResults('skills'), getSkills);
 router.route('/:id').get(getSkill);
