+++ routes/api/user-routes.js
@@ -1,16 +1,16 @@
 const express = require('express');
 const router = express.Router();
 const { check, oneOf } = require('express-validator');
-const advancedResults = require('../middleware/advancedResults');
-const { protect, authorise } = require('../middleware/auth');
-const { checkInputError } = require('../middleware/inputValidation');
-const { ALPHA_WHITESPACE_REGEX } = require('../utils/regex');
+const advancedResults = require('../../middleware/advancedResults');
+const { protect, authorise } = require('../../middleware/auth');
+const { checkInputError } = require('../../middleware/inputValidation');
+const { ALPHA_WHITESPACE_REGEX } = require('../../utils/regex');
 const {
   INVALID_EMAIL_MSG,
   INVALID_ALPHA_SPACE_MSG,
   INVALID_PASSWORD_MSG,
   NO_FIELD_UPDATED_MSG,
-} = require('../utils/inputExceptionMsg');
+} = require('../../utils/inputExceptionMsg');
 
 // import controllers here
 const {
@@ -19,7 +19,7 @@ const {
   createUser,
   updateUser,
   deleteUser,
-} = require('../controllers/users');
+} = require('../../controllers/users');
 
 // all route to use protect middleware
 router.use(protect);
